Update formatting on organisation phone numbers in requests to Serco
